api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "API for eh global hack"
authors = ["Prasanjit Prakash <prasanjit@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.2"
flask-sqlalchemy = "^3.0.5"
psycopg = "^3.1.10"
python-dotenv = "^1.0.0"
loguru = "^0.7.0"
attrs = "^23.1.0"
flask-cors = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
